requirements.txt
flask==2.0.1
torch==1.9.0
torchvision==0.10.0
pillow==8.3.1
efficientnet-pytorch==0.7.1
gunicorn==20.1.0